==> 03_nlp_glove_features.ipynb <==
<h2> 3.6 Featurizing text data with tfidf weighted word-vectors </h2>

import pandas as pd
import matplotlib.pyplot as plt
import re
import time
import warnings
import numpy as np
from nltk.corpus import stopwords
from sklearn.preprocessing import normalize
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
warnings.filterwarnings("ignore")
import sys
import os 
import pandas as pd
import numpy as np
from tqdm import tqdm
import joblib

# exctract word2vec vectors
# https://github.com/explosion/spaCy/issues/1721
# http://landinghub.visualstudio.com/visual-cpp-build-tools
import spacy

data_points = 3000

# avoid decoding problems
df_data = pd.read_csv("data/00_train.csv")

df_data['question1'] = df_data['question1'].apply(lambda x: str(x))
df_data['question2'] = df_data['question2'].apply(lambda x: str(x))
df_data.head(1)

## Checkpoint 1: Applying TFIDF WEIGHT W2V

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer

questions = list(df_data['question1'][:data_points] + df_data['question2'][:data_points])

tfidf = TfidfVectorizer(lowercase=False, )
tfidf.fit_transform(questions)

# dict key:word and value:tf-idf score
word2tfidf = dict(zip(tfidf.get_feature_names(), tfidf.idf_))

counter = 0
for k, v in word2tfidf.items():
    if counter == 4:
        break
    else:
        counter += 1
        print(k, ' ', v)

- After we find TF-IDF scores, we convert each question to a weighted average of word2vec vectors by these scores.
- here we use a pre-trained GLOVE model which comes free with "Spacy".  https://spacy.io/usage/vectors-similarity
- It is trained on Wikipedia and therefore, it is stronger in terms of word semantics. 

# en_vectors_web_lg, which includes over 1 million unique vectors.
nlp = spacy.load('en_core_web_sm')

def tifidf_to_wieghtedW2V(text_df, tfidf_vectors):
    vecs = []
    # https://github.com/noamraph/tqdm
    # tqdm is used to print the progress bar
    for ques in tqdm(list(text_df)):
        doc = nlp(ques) 
        # 384 is the number of dimensions of vectors 
        mean_vec = np.zeros([len(doc), len(doc[0].vector)])
        for word in doc:
            # word2vec
            vec = word.vector
            # fetch df score
            try:
                idf = tfidf_vectors[str(word)]
            except:
                idf = 0
            # compute final vec
            mean_vec += vec * idf
        mean_vec = mean_vec.mean(axis=0)
        vecs.append(mean_vec)
    return vecs

q1 = list(tifidf_to_wieghtedW2V(df_data['question1'][:data_points], word2tfidf))

joblib.dump(q1, f'models/01_nlp/03_1_tfidf_weighted_w2v_{data_points}.joblib')

q2 = list(tifidf_to_wieghtedW2V(df_data['question2'][:data_points], word2tfidf))

joblib.dump(q2, f'models/01_nlp/04_2_tfidf_weighted_w2v{data_points}.joblib')

## Checkpoint 3: Applying Glove

from gensim.models import Word2Vec
from gensim.models import KeyedVectors
import pickle
import gensim

# from gensim.scripts.glove2word2vec import glove2word2vec
# glove2word2vec(glove_input_file="data/glove.42B.300d.txt", word2vec_output_file="data/glove_vectors.txt")

# !python -m gensim.scripts.glove2word2vec --input  data/glove.840B.300d.txt --output data/glove.840B.300d.w2vformat.txt

from gensim.models.keyedvectors import KeyedVectors
glove_model = KeyedVectors.load_word2vec_format("data/glove_vectors.txt", binary=False, unicode_errors='ignore')

def wmd(s1, s2,model):
    s1 = str(s1)
    s2 = str(s2)
    s1 = s1.split()
    s2 = s2.split()
    return model.wmdistance(s1, s2)

http://proceedings.mlr.press/v37/kusnerb15.pdf i read about word mover distance and after that i calculated some distances from avg word vectors as below

df = df_data 
df['Word_Mover_Dist'] = df.apply(lambda x: wmd(x['question1'], x['question2'],glove_model), axis=1)

# the avg-w2v for each sentence/review is stored in this list
def avg_w2v(list_of_sent,model,d):
    '''
    Returns average of word vectors for
    each sentance with dimension of model given
    '''
    sent_vectors = []
    for sent in list_of_sent: # for each review/sentence
        doc = [word for word in sent if word in model.key_to_index]
        if doc: 
            sent_vec = np.mean(model[doc],axis=0)
        else:
            sent_vec = np.zeros(d)
        sent_vectors.append(sent_vec)
    return sent_vectors

#converting into lists
list_of_question1=[]
for sent in df.question1.values:
    list_of_question1.append(sent.split())
list_of_question2=[]
for sent in df.question2.values:
    list_of_question2.append(sent.split())

#avg word 2 vec
# d= [word for word in sent if word in glove_model.key_to_index]
# sent_vec = np.mean(glove_model[d],axis=0)

avgw2v_q1 = avg_w2v(list_of_question1,glove_model,300)
avgw2v_q2 = avg_w2v(list_of_question2,glove_model,300)

#converting as df
df_avgw2v = pd.DataFrame()
df_avgw2v['q1_vec'] = list(avgw2v_q1)
df_avgw2v['q2_vec'] = list(avgw2v_q2)
df_q1 = pd.DataFrame(df_avgw2v.q1_vec.values.tolist())
df_q2 = pd.DataFrame(df_avgw2v.q2_vec.values.tolist())

df_q1.head()

#importing soma distances and calculating
from scipy.stats import skew, kurtosis
from scipy.spatial.distance import cosine, cityblock,canberra, euclidean, minkowski
from scipy.spatial.distance import braycurtis, chebyshev, correlation, mahalanobis
from scipy.spatial.distance import seuclidean, hamming, jaccard, kulsinski, rogerstanimoto, russellrao, sokalmichener, sokalsneath, kulsinski, yule

df['dist_cosine'] = [cosine(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_cityblock'] = [cityblock(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_canberra'] = [canberra(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_euclidean'] = [euclidean(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_minkowski'] = [minkowski(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_braycurtis'] = [braycurtis(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_chebyshev'] = [chebyshev(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_correlation'] = [correlation(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_hamming'] = [hamming(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_jaccard'] = [jaccard(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_kulsinski'] = [kulsinski(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_rogerstanimoto'] = [rogerstanimoto(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_russellrao'] = [russellrao(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_sokalmichener'] = [sokalmichener(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_kulsinski'] = [kulsinski(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]
df['dist_yule'] = [yule(x, y) for (x, y) in zip(avgw2v_q1,avgw2v_q2)]

#filling na values with 0  for cosine distance
df = df.fillna(0)

print(df.shape)
df.head(1)

df.to_csv("data/03_nlp_glove_features.csv")

question1 = df_data['question1'][:2000]

tfidf_vect = TfidfVectorizer(ngram_range=(1,2), min_df=10, lowercase=False)
tfidf_1 = tfidf_vect.fit_transform(question1)

joblib.dump(tfidf_1, "models/01_nlp/01_1_tfidf_model.joblib")

question2 = df_data['question2'][:2000]

tfidf_vect = TfidfVectorizer(ngram_range=(1,2), min_df=10, lowercase=False)
tfidf_2 = tfidf_vect.fit_transform(question2)
type(tfidf_2)

joblib.dump(tfidf_2, "models/01_nlp/02_2_tfidf_model.joblib")

tfidf_vect = TfidfVectorizer(ngram_range=(1,3),max_features=200000,min_df=0.000032)
train_tfidf = tfidf_vect.fit_transform(X_train_tf.Text)
test_tfidf = tfidf_vect.transform(X_test_tf.Text)
print('No of Tfidf features',len(tfidf_vect.get_feature_names()))

from scipy.sparse import hstack
X_train1 = hstack((X_train_tf.values,train_tfidf))
X_test1 = hstack((X_test_tf.values,test_tfidf))

scale = StandardScaler(with_mean=False)
X_train_sc = scale.fit_transform(X_train1)
X_test_sc = scale.transform(X_test1)
