==> 04_features_merging.ipynb <==
import pandas as pd
import scipy.sparse
import joblib 

# Creating db file from csv
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

from sqlalchemy import create_engine # database connection
import sqlite3

from sklearn.model_selection import GridSearchCV

import os
import datetime as dt
from collections import Counter, defaultdict

from xgboost import XGBClassifier

from sklearn.model_selection import train_test_split

from sklearn.metrics import accuracy_score, log_loss, confusion_matrix

from sklearn.calibration import CalibratedClassifierCV
from sklearn.linear_model import SGDClassifier

data_points = 3000

mat = pd.read_csv("data/00_train.csv")
mat = mat.drop(['qid1','qid2','question1','question2',],axis=1)
# df_y = mat[:data_points]
df_y = mat
df_y.head(1)

## Checkpoint 1: Merging TFIDF WEIGHTED W2V

mat = joblib.load('models/01_nlp/03_1_tfidf_weighted_w2v.joblib')
tfidf_weighted_w2v_1 = pd.DataFrame(mat)
tfidf_weighted_w2v_1.head(1)

mat = joblib.load('models/01_nlp/04_2_tfidf_weighted_w2v.joblib')
tfidf_weighted_w2v_2 = pd.DataFrame(mat)
tfidf_weighted_w2v_2.head(1)

tfidf_weighted_w2v_1['id']= df_y['id']
tfidf_weighted_w2v_2['id']= df_y['id']

tfidf_weighted_w2v_full = tfidf_weighted_w2v_1.merge(tfidf_weighted_w2v_2, on='id',how='left')
tfidf_weighted_w2v_full.head(2)

joblib.dump(tfidf_weighted_w2v_full, f"models/01_nlp/06_tfidf_weighted_w2v_{data_points}_data_points.joblib")

## Checkpoint 2: Merging TFIDF

mat = joblib.load('models/01_nlp/01_1_tfidf_model.joblib') # sparse
mat1 = pd.DataFrame.sparse.from_spmatrix(mat)

mat1.head(1)
# tfidf_1_dense  = mat.sparse.to_dense() # dense
# tfidf_1_dense.head(1)

mat = joblib.load('models/01_nlp/02_2_tfidf_model.joblib') # sparse
mat2 = pd.DataFrame.sparse.from_spmatrix(mat)

mat2.head(1)
# tfidf_2_dense  = mat.sparse.to_dense() # dense
# tfidf_2_dense.head(1)

mat1['id'] = df_y['id']
mat2['id'] = df_y['id']


df_tfidf_full = mat1.merge(mat2, on='id',how='left')
df_tfidf_full.head(2)

# tfidf_1_dense['id'] = df_y['id']
# tfidf_2_dense['id'] = df_y['id']

# df_tfidf_full = tfidf_1_dense.merge(tfidf_2_dense, on='id',how='left')
# df_tfidf_full.head(2)

# joblib.dump(df_tfidf_full, f"models/01_nlp/05_full_tfidf_model_{data_points}_datapoints.joblib")

# joblib.dump(df_tfidf_full, f"models/01_nlp/05_full_tfidf_model_{data_points}datapoints.joblib")

## Checkpoint 3: Merging basic, advanced & nlp distance features using eda

mat = pd.read_csv("data/00_train.csv")
# mat = mat.drop(['qid1','qid2','question1','question2',],axis=1)
# df_y = mat[:data_points]
df_y = mat
df_y.head(1)

import re
import numpy as np
import pandas as pd

from nltk.corpus import stopwords
from nltk.stem import PorterStemmer

from bs4 import BeautifulSoup

# To get the results in 4 decemal points
SAFE_DIV = 0.0001 

STOP_WORDS = stopwords.words("english")


def preprocess(x):
    x = str(x).lower()
    x = x.replace(",000,000", "m").replace(",000", "k").replace("′", "'").replace("’", "'")\
                           .replace("won't", "will not").replace("cannot", "can not").replace("can't", "can not")\
                           .replace("n't", " not").replace("what's", "what is").replace("it's", "it is")\
                           .replace("'ve", " have").replace("i'm", "i am").replace("'re", " are")\
                           .replace("he's", "he is").replace("she's", "she is").replace("'s", " own")\
                           .replace("%", " percent ").replace("₹", " rupee ").replace("$", " dollar ")\
                           .replace("€", " euro ").replace("'ll", " will")
    x = re.sub(r"([0-9]+)000000", r"\1m", x)
    x = re.sub(r"([0-9]+)000", r"\1k", x)
    
    
    porter = PorterStemmer()
    pattern = re.compile('\W')
    
    if type(x) == type(''):
        x = re.sub(pattern, ' ', x)
    
    if type(x) == type(''):
        x = porter.stem(x)
        example1 = BeautifulSoup(x)
        x = example1.get_text()
               
    return x

df_y['cleaned_q1'] = preprocess(df_y['question1'][0])
df_y['cleaned_q2'] = preprocess(df_y['question2'][0])

df_basic = pd.read_csv("data/01_basic_eda.csv")
df_basic = df_basic.drop(['qid1','qid2','question1','question2','is_duplicate'],axis=1)
# df_basic = df_basic
df_basic.head(1)

df_advance = pd.read_csv("data/02_advanced_eda.csv")
df_advance = df_advance.drop(['qid1','qid2','question1','question2','is_duplicate'],axis=1)
# df_advance = df_advance
df_advance.head(1)

df_nlp_distance = pd.read_csv("data/03_nlp_glove_features.csv")
df_nlp_distance = df_nlp_distance.drop(['Unnamed: 0', 'qid1','qid2','question1','question2','is_duplicate'],axis=1)
# df_advance = df_advance
df_nlp_distance.head(1)

df_extraction = df_y.merge(df_basic, on='id',how='left')
df_extraction = df_extraction.merge(df_advance, on='id',how='left')
df_extraction = df_extraction.merge(df_nlp_distance, on='id',how='left')
# df_extraction = df_extraction.drop(['id','is_duplicate'],axis=1)
# df_extraction = df_extraction.drop(['id'],axis=1)
print(df_extraction.shape)
df_extraction.head(1)

df_extraction.to_csv(f"data/04_all_features.csv")
